[#18] Refactor: 로그인 모달 창 LoginModal 컴포넌트로 분리
